+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Fold effect of panels.
+- Skip button for mobile.
+- GA code.
+
 ## [0.2] - 2017-02-08
 ### Added
 - Search form.